refactor: remove the `entities` folder from the game core
