+++ src/game-core/business/typings/business.typings.ts
