+++ src/game-core/business/utils/business.utils.ts
