+++ src/game-core/player/typings/index.ts
