+++ src/game-core/player/typings/player.typings.ts
